+++ game/game_libs/ui/menus/ConnectionProgress.cpp
@@ -434,7 +434,7 @@ void UI_ConnectionProgress_f(void)
 	}
 	else if ( !strcmp(EngFuncs::CmdArgv(1), "stufftext") )
 	{
-		menu_connectionprogress->HandleStufftext(atof(EngFuncs::CmdArgv(2)), EngFuncs::CmdArgv(3));
+		menu_connectionprogress->HandleStufftext(static_cast<float>(atof(EngFuncs::CmdArgv(2))), EngFuncs::CmdArgv(3));
 	}
 	else if ( !strcmp(EngFuncs::CmdArgv(1), "precache") )
 	{
