+++ game/game_libs/ui/menus/VideoModes.cpp
@@ -166,7 +166,7 @@ void CMenuVidModes::SetConfig()
 	if ( isVidModeChanged )
 	{
 		SetMode(currentModeIndex);
-		EngFuncs::CvarSetValue("vid_mode", currentModeIndex);
+		EngFuncs::CvarSetValue("vid_mode", static_cast<float>(currentModeIndex));
 		vidList.SetCurrentIndex(currentModeIndex + VID_MODES_POS);
 		// have changed resolution, but enable test mode only in fullscreen
 		testMode |= !windowed.bChecked;
@@ -183,10 +183,10 @@ void CMenuVidModes::SetConfig()
 
 void CMenuVidModes::ApplyChanges()
 {
-	prevMode = EngFuncs::GetCvarFloat("vid_mode");
+	prevMode = static_cast<int>(EngFuncs::GetCvarFloat("vid_mode"));
 	prevFullscreen = EngFuncs::GetCvarFloat("fullscreen");
-	prevModeX = EngFuncs::GetCvarFloat("width");
-	prevModeY = EngFuncs::GetCvarFloat("height");
+	prevModeX = static_cast<int>(EngFuncs::GetCvarFloat("width"));
+	prevModeY = static_cast<int>(EngFuncs::GetCvarFloat("height"));
 }
 
 void CMenuVidModes::RevertChanges()
