+++ game/game_libs/ui/menus/botsetup/BotProfileListModel.cpp
@@ -67,7 +67,7 @@ void CBotProfileListModel::InitialiseProfileTable()
 		return;
 	}
 
-	m_IndexToProfileName.EnsureCapacity(m_ProfileTable.Count());
+	m_IndexToProfileName.EnsureCapacity(static_cast<int>(m_ProfileTable.Count()));
 
 	for ( CBotProfileTable::ConstIterator it = m_ProfileTable.CBegin(); it != m_ProfileTable.CEnd(); ++it )
 	{
